Created new application init process and convert components to modules.

+++ jujugui/static/gui/src/app/init.js
@@ -1,31 +1,37 @@
-// /* Copyright (C) 2017 Canonical Ltd. */
+/* Copyright (C) 2017 Canonical Ltd. */
 'use strict';
 
-/*
-  In an effort to develop this init.js along side with the rest of the
-  app here are the changes you'll have to manually change to continue
-  working on the new system.
-
-  index.html.mako
-    - Uncomment the init-pkg.js script tags lines 300, 308.
-    - Uncomment the initialization code lines 353, 354.
-    - Comment the old init code lines 352, 357.
-  Makefile
-    - Uncomment the hack to generate the init-pkg file Line 221.
-
-  Code to still move over from app.js
-    - Y.juju.Cookies Line 50.
-    - widgets.AutodeployExtension Line 49.
-*/
-
+const React = require('react');
+const ReactDOM = require('react-dom');
 const mixwith = require('mixwith');
 
 const utils = require('./init/utils');
+const viewUtils = require('./views/utils');
 const hotkeys = require('./init/hotkeys');
 const csUser = require('./init/charmstore-user');
+const cookieUtil = require('./init/cookie-util');
+const BundleImporter = require('./init/bundle-importer');
+
+const newBakery = require('./init/utils/bakery-utils');
 
 const ComponentRenderersMixin = require('./init/component-renderers-mixin');
 const DeployerMixin = require('./init/deployer-mixin');
+
+// Hacks untill all of the global references have been removed.
+window.jsyaml = require('js-yaml');
+window.juju.utils.RelationUtils = require('./init/relation-utils');
+// Required for the envionment.js file.
+window.ReactDOM = ReactDOM;
+window.React = React;
+juju.components.AmbiguousRelationMenu = require(
+  './components/relation-menu/ambiguous-relation-menu');
+juju.components.Environment = require(
+  './components/environment/environment');
+juju.components.RelationMenu = require(
+  './components/relation-menu/relation-menu');
+window.views = {
+  PanZoomModule: require('./init/topology/panzoom.js')
+};
 
 const yui = window.yui;
 
@@ -43,7 +49,7 @@ class GUIApp {
       portion of the application has set this value prior to switching models.
       @type {String}
     */
-    this.modelUUID = config.jujuEnvUUID;
+    this.modelUUID = config.jujuEnvUUID || null;
     /**
       The default web page title.
       @type {String}
@@ -64,16 +70,6 @@ class GUIApp {
     */
     this.anonymousMode = false;
     /**
-      Reference to the changesUtils utilities.
-      @type {Object}
-    */
-    this.changesUtils = window.juju.utils.ChangesUtils;
-    /**
-      Reference to the relationUtils utilities.
-      @type {Object}
-    */
-    this.relationUtils = window.juju.utils.RelationUtils;
-    /**
       Stores the custom event handlers for the application.
       @type {Object}
     */
@@ -83,11 +79,6 @@ class GUIApp {
       @type {Integer}
     */
     this._dragLeaveTimer = null;
-    /**
-      Reference to the juju.Cookies instance.
-      @type {juju.Cookies}
-    */
-    this._cookieHandler = null;
     /**
       The keydown event listener from the hotkey activation.
       @type {Object}
@@ -131,7 +122,7 @@ class GUIApp {
       Used to perform requests on a macaroon authenticated endpoints.
       @type {Object}
     */
-    this.bakery = yui.juju.bakeryutils.newBakery(
+    this.bakery = newBakery(
       config, this.user, stateGetter, cookieSetter, webHandler);
     /**
       A charm store API client instance.
@@ -148,7 +139,6 @@ class GUIApp {
       config, window.jujulib.bundleservice);
 
     this.ecs = this._setupEnvironmentChangeSet();
-
     const modelOptions = {
       user: this.user,
       ecs: this.ecs,
@@ -249,7 +239,7 @@ class GUIApp {
       Application instance of the bundle importer.
       @type {Object}
     */
-    this.bundleImporter = new yui.juju.BundleImporter({
+    this.bundleImporter = new BundleImporter({
       db: this.db,
       modelAPI: this.modelAPI,
       getBundleChanges: this.controllerAPI.getBundleChanges.bind(
@@ -271,7 +261,7 @@ class GUIApp {
     */
     this.acl = new yui.juju.generateAcl(this.controllerAPI, this.modelAPI);
     // Listen for window unloads and trigger the unloadWindow function.
-    window.onbeforeunload = yui.juju.views.utils.unloadWindow.bind(this);
+    window.onbeforeunload = utils.unloadWindow.bind(this);
 
     this._handleMaasServer();
     // Feed environment changes directly into the database.
@@ -319,7 +309,10 @@ class GUIApp {
     */
     this.topology = this._renderTopology();
     this._renderComponents();
-    this.state.bootstrap();
+    const result = this.state.bootstrap();
+    if (result.error) {
+      console.error(result.error);
+    }
   }
 
   /**
@@ -341,8 +334,8 @@ class GUIApp {
     };
     // Bind switchModel separately to include the already bound
     // addNotifications.
-    this._bound.switchModel = yui.juju.views.utils.switchModel.bind(
-      this, this.env, this._bound.addNotification);
+    this._bound.switchModel = utils.switchModel.bind(
+      this, this.modelAPI, this._bound.addNotification);
   }
 
   /**
@@ -427,10 +420,8 @@ class GUIApp {
     @param {Function} next - The next route handler.
   */
   authorizeCookieUse(state, next) {
-    var GTM_enabled = this.GTM_enabled;
-    if (GTM_enabled) {
-      this._cookieHandler = this._cookieHandler || new yui.juju.Cookies();
-      this._cookieHandler.check();
+    if (this.applicationConfig.GTM_enabled) {
+      cookieUtil.check(document);
     }
     next();
   }
@@ -489,6 +480,7 @@ class GUIApp {
       ['*', this._ensureControllerConnection.bind(this)],
       ['*', this.authorizeCookieUse.bind(this)],
       ['*', this.checkUserCredentials.bind(this)],
+      ['*', this._renderComponents.bind(this)],
       ['root',
         this._rootDispatcher.bind(this),
         this._clearRoot.bind(this)],
@@ -526,6 +518,9 @@ class GUIApp {
       ['gui.deploy',
         this._renderDeployment.bind(this),
         this._clearDeployment.bind(this)],
+      ['postDeploymentPanel',
+        this._displayPostDeployment.bind(this),
+        this._clearPostDeployment.bind(this)],
       // Nothing needs to be done at the top level when the hash changes.
       ['hash'],
       // special dd is handled by the root dispatcher as it requires /new
@@ -783,15 +778,6 @@ class GUIApp {
       }));
   }
 
-  /**
-    Hide the drag notifications.
-  */
-  _hideDragOverNotification() {
-    this.topology.fadeHelpIndicator(false);
-    ReactDOM.unmountComponentAtNode(
-      document.getElementById('drag-over-notification-container'));
-  }
-
   _controllerLoginHandler(entityPromise, evt) {
     const state = this.state;
     this.anonymousMode = false;
@@ -878,7 +864,7 @@ class GUIApp {
         !newState.profile &&
         newState.root !== 'account' &&
         (isLogin || !current.root) &&
-        this.gisf
+        this.applicationConfig.gisf
       ) {
         newState.profile = this.user.displayName;
       }
@@ -1121,7 +1107,7 @@ class GUIApp {
     if (!err) {
       return;
     }
-    if (!yui.juju.views.utils.isRedirectError(err)) {
+    if (!viewUtils.isRedirectError(err)) {
       // There is nothing to do in this case, and the user is already
       // prompted with the error in the login view.
       console.log(`cannot log into ${api.name}: ${err}`);
@@ -1164,7 +1150,7 @@ class GUIApp {
           uuid: this.modelUUID,
           server: publicHost.value,
           port: publicHost.port
-        }, null, null, null, true, false));
+        }), null, null, null, true, false);
     });
   }
 
@@ -1473,6 +1459,9 @@ class GUIApp {
     this.state.changeState({
       gui: {
         deploy: JSON.stringify(ddData)
+      },
+      postDeploymentPanel: {
+        entityId: ddData.id
       }
     });
   }
@@ -1622,14 +1611,15 @@ class GUIApp {
       clearTimeout(this._dbChangedTimer);
     }
     // Destroy YUI classes.
-    this.modelAPI.destroy();
+    this.modelAPI && this.modelAPI.destroy();
     this.controllerAPI.destroy();
     this.db.destroy();
     this.endpointsController.destroy();
+    this.topology.destroy();
     // Detach event listeners.
-    const remove = document.removeEventListener;
+    const remove = document.removeEventListener.bind(document);
     const handlers = this._domEventHandlers;
-    this._hotkeyListener.detach();
+    this._hotkeyListener.deactivate();
     const ecsListener = handlers['renderDeploymentBarListener'];
     remove('ecs.changeSetModified', ecsListener);
     remove('ecs.currentCommitFinished', ecsListener);
